Rename mutation webhook
